ai-engine/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.2.0

# PDF Parsing
pymupdf>=1.23.0
pdfplumber>=0.10.0

# Google Gemini
google-genai>=1.0.0

# RAG / Vector Store
chromadb>=0.5.0
sentence-transformers>=3.0.0

# LangChain
langchain>=0.2.0
langchain-community>=0.2.0

# Pinecone (optional fallback)
pinecone-client>=3.0.0

# HTTP
httpx>=0.27.0
aiohttp>=3.9.0
requests>=2.31.0

# Real-Time Job Search (SerpApi)
google-search-results>=2.4.2

# Utils
python-dotenv>=1.0.0
numpy>=1.26.0
scipy>=1.12.0
groq>=1.1.0
